Address feedback from PR first round
